requirements.txt
# Core dependencies
langchain==1.0.2
langchain-core==1.0.1
langchain-community==0.0.4
langchain-ollama==1.0.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# UI and visualization
streamlit>=1.29.0
plotly>=5.18.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0

# Note: Ollama should be installed separately from https://ollama.ai